requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
geopy>=2.2.0
aiohttp>=3.8.1
python-dotenv>=0.19.0
crewai>=0.1.0
langchain>=0.0.150